+++ examples/alpha_with_custom_datasource.py
@@ -4,7 +4,7 @@ import asyncio
 import logging
 import polars as pl
 from typing import override, cast
-from datetime import datetime
+from datetime import datetime, timedelta
 
 from pathlib import Path
 from adrs import Alpha, DataLoader
@@ -234,7 +234,7 @@ async def main():
         start_time=start_time,
         end_time=end_time,
         fees=fees,
-        price_shift=10,  # assume 10 minutes delay
+        execution_delay=timedelta(minutes=10),  # execution delay after bar close
     )
     print(performance)
     print(df)
@@ -263,7 +263,7 @@ async def main():
         datamap=datamap,
         data_df=data_df,
         fees=fees,
-        price_shift=10,  # assume 10 minutes delay
+        execution_delay=timedelta(minutes=10),  # execution delay after bar close
     )
     print("backtest", report.back.sensitivity_sr_summary)
     print("forward test", report.forward.sensitivity_sr_summary)